# back/src/main/java/com/example/controller/NewsStatisticsController.java
package com.example.controller;

import java.util.List;
import jakarta.servlet.http.HttpServletResponse;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import com.ruoyi.common.annotation.Log;
import com.ruoyi.common.core.controller.BaseController;
import com.ruoyi.common.core.domain.AjaxResult;
import com.ruoyi.common.enums.BusinessType;
import com.example.domain.NewsStatistics;
import com.example.service.INewsStatisticsService;
import com.ruoyi.common.utils.poi.ExcelUtil;
import com.ruoyi.common.core.page.TableDataInfo;

/**
 * 新闻统计Controller
 * 
 * @author tong
 * @date 2026-05-09
 */
@RestController
@RequestMapping("/system/statistics")
public class NewsStatisticsController extends BaseController
{
    @Autowired
    private INewsStatisticsService newsStatisticsService;

    /**
     * 查询新闻统计列表
     */
    @PreAuthorize("@ss.hasPermi('system:statistics:list')")
    @GetMapping("/list")
    public TableDataInfo list(NewsStatistics newsStatistics)
    {
        startPage();
        List<NewsStatistics> list = newsStatisticsService.selectNewsStatisticsList(newsStatistics);
        return getDataTable(list);
    }

    /**
     * 导出新闻统计列表
     */
    @PreAuthorize("@ss.hasPermi('system:statistics:export')")
    @Log(title = "新闻统计", businessType = BusinessType.EXPORT)
    @PostMapping("/export")
    public void export(HttpServletResponse response, NewsStatistics newsStatistics)
    {
        List<NewsStatistics> list = newsStatisticsService.selectNewsStatisticsList(newsStatistics);
        ExcelUtil<NewsStatistics> util = new ExcelUtil<NewsStatistics>(NewsStatistics.class);
        util.exportExcel(response, list, "新闻统计数据");
    }

    /**
     * 获取新闻统计详细信息
     */
    @PreAuthorize("@ss.hasPermi('system:statistics:query')")
    @GetMapping(value = "/{id}")
    public AjaxResult getInfo(@PathVariable("id") Long id)
    {
        return success(newsStatisticsService.selectNewsStatisticsById(id));
    }

    /**
     * 新增新闻统计
     */
    @PreAuthorize("@ss.hasPermi('system:statistics:add')")
    @Log(title = "新闻统计", businessType = BusinessType.INSERT)
    @PostMapping
    public AjaxResult add(@RequestBody NewsStatistics newsStatistics)
    {
        return toAjax(newsStatisticsService.insertNewsStatistics(newsStatistics));
    }

    /**
     * 修改新闻统计
     */
    @PreAuthorize("@ss.hasPermi('system:statistics:edit')")
    @Log(title = "新闻统计", businessType = BusinessType.UPDATE)
    @PutMapping
    public AjaxResult edit(@RequestBody NewsStatistics newsStatistics)
    {
        return toAjax(newsStatisticsService.updateNewsStatistics(newsStatistics));
    }

    /**
     * 删除新闻统计
     */
    @PreAuthorize("@ss.hasPermi('system:statistics:remove')")
    @Log(title = "新闻统计", businessType = BusinessType.DELETE)
	@DeleteMapping("/{ids}")
    public AjaxResult remove(@PathVariable Long[] ids)
    {
        return toAjax(newsStatisticsService.deleteNewsStatisticsByIds(ids));
    }
}